Encapsulate state. Expose behind interface for easier test mocking
